chapter-05/01-data-enrichment/01-static-joiner-data-in-motion/requirements.txt
delta-spark==3.0.0
pyspark==3.5.0
